Feature/retweet

See merge request AlirezaRoshanzamir/bootcamp!4

+++ bootcamp/feeds/views.py
@@ -133,16 +133,16 @@ def post(request):
 @ajax_required
 def like(request):
     feed_id = request.POST['feed']
-    feed = Feed.objects.get(pk=feed_id)
+    feed = Feed.objects.get(pk=feed_id).get_source_feed()
     user = request.user
-    like = Activity.objects.filter(activity_type=Activity.LIKE, feed=feed_id,
+    like = Activity.objects.filter(activity_type=Activity.LIKE, feed=feed.id,
                                    user=user)
     if like:
         user.profile.unotify_liked(feed)
         like.delete()
 
     else:
-        like = Activity(activity_type=Activity.LIKE, feed=feed_id, user=user)
+        like = Activity(activity_type=Activity.LIKE, feed=feed.id, user=user)
         like.save()
         user.profile.notify_liked(feed)
 
@@ -154,7 +154,7 @@ def like(request):
 def comment(request):
     if request.method == 'POST':
         feed_id = request.POST['feed']
-        feed = Feed.objects.get(pk=feed_id)
+        feed = Feed.objects.get(pk=feed_id).get_source_feed()
         post = request.POST['post']
         post = post.strip()
         if len(post) > 0:
@@ -169,7 +169,7 @@ def comment(request):
 
     else:
         feed_id = request.GET.get('feed')
-        feed = Feed.objects.get(pk=feed_id)
+        feed = Feed.objects.get(pk=feed_id).get_source_feed()
         return render(request, 'feeds/partial_feed_comments.html',
                       {'feed': feed})
 
@@ -186,7 +186,7 @@ def update(request):
 
     dump = {}
     for feed in feeds:
-        dump[feed.pk] = {'likes': feed.likes, 'comments': feed.comments}
+        dump[feed.pk] = {'likes': feed.get_likes_count(), 'comments': feed.get_comments_count()}
 
     data = json.dumps(dump)
     return HttpResponse(data, content_type='application/json')
@@ -196,7 +196,7 @@ def update(request):
 @ajax_required
 def track_comments(request):
     feed_id = request.GET.get('feed')
-    feed = Feed.objects.get(pk=feed_id)
+    feed = Feed.objects.get(pk=feed_id).get_source_feed()
     if len(feed.get_comments()) > 0:
         return render(
             request, 'feeds/partial_feed_comments.html', {'feed': feed})
@@ -212,10 +212,11 @@ def remove(request):
         feed_id = request.POST.get('feed')
         feed = Feed.objects.get(pk=feed_id)
         if feed.user == request.user:
-            likes = feed.get_likes()
             parent = feed.parent
-            for like in likes:
-                like.delete()
+            if not feed.is_retweet():
+                likes = feed.get_likes()
+                for like in likes:
+                    like.delete()
 
             feed.delete()
             if parent:
@@ -228,3 +229,22 @@ def remove(request):
 
     except Exception:
         return HttpResponseBadRequest()
+
+@login_required
+@ajax_required
+def retweet(request):
+    last_feed = request.POST.get('last_feed')
+    source_feed_id = request.POST.get('feed')
+    csrf_token = (csrf(request)['csrf_token'])
+    user = request.user
+
+    source_feed = Feed.objects.get(pk=source_feed_id)
+    source_feed.retweet(user)
+
+    html = _html_feeds(last_feed, user, csrf_token)
+    return HttpResponse(html)
+
+@login_required
+@ajax_required
+def remove_retweet(request):
+    return remove(request)
